requirements.txt
googlemaps==2.5.1
Pillow==4.0.0

# Run these in a clone of https://github.com/hzeller/rpi-rgb-led-matrix
# make build-python
# make install-python  # Note: no `sudo`, like in the official docs
